nodes/src/test/java/de/eldoria/bloodnight/nodes/transform/impl/logical/NotNodeTest.java
package de.eldoria.bloodnight.nodes.transform.impl.logical;

import de.eldoria.bloodnight.nodes.base.io.Edge;
import de.eldoria.bloodnight.nodes.container.NodeContainer;
import de.eldoria.bloodnight.nodes.meta.DataType;
import de.eldoria.bloodnight.nodes.meta.Fields;
import de.eldoria.bloodnight.nodes.value.impl.BooleanNode;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.CsvSource;

import static org.junit.jupiter.api.Assertions.assertEquals;

class NotNodeTest {
    @ParameterizedTest
    @CsvSource({
            "true,false",
            "false,true"
    })
    void output(boolean first, boolean expectedResult) {
        NodeContainer nodeContainer = new NodeContainer();
        nodeContainer.add(1, new BooleanNode(first));
        var notNode = nodeContainer.add(2, new NotNode());
        notNode.input().connect(Fields.FIRST, new Edge(1, Fields.VALUE));
        var output = notNode.output();
        assertEquals(expectedResult, output.value(Fields.RESULT));
        assertEquals(DataType.BOOLEAN, output.getType(Fields.RESULT));
    }
}
